Runs all functions without errors but does not work

+++ toolbox.cs
@@ -18,7 +18,8 @@ namespace sharpclean
         //gets some info for saving data, then taps run()
         public void clean()
         {
-            if (pixels == null) {
+            if (pixels == null)
+            {
                 Console.WriteLine(toolbox_err + "no pixels loaded\n");
                 return;
             }
@@ -26,10 +27,10 @@ namespace sharpclean
             int n = 0;
             ofilename = "none";
 
-            if (cmd.getcmd("write data to .csv file? [1]yes, [2]no, [q]quit - ", ref n, 2)) {
+            if (cmd.getcmd("write data to .csv file? [1]yes, [2]no, [q]quit - ", ref n, 2))
+            {
                 if (n == 1)
                     cmd.getfile("enter data output file name : ", ref ofilename, ".csv", 2);
-
                 run();
             }
         }
@@ -38,22 +39,18 @@ namespace sharpclean
         private void run()
         {
             System.Diagnostics.Stopwatch watch = new System.Diagnostics.Stopwatch();
-
             int per_25 = totalPixels / 4;
             int per_50 = totalPixels / 2;
             int per_75 = per_25 + per_50;
-
             bool b_25 = false, b_50 = false, b_75 = false;
-
             bool writeData = false;
-            if (ofilename != "none") {
+            if (ofilename != "none")
+            {
                 ofilename = "data/" + ofilename;
                 System.IO.File.WriteAllText(ofilename, "val, size, edge, dust, obj, res, type, c avg, c edge, c size\n");
                 writeData = true;
             }
-
             selection s = new selection(pixels, imageWidth, totalPixels);
-
             watch.Start();
             for (int i = 0; i < totalPixels; i++)
             {
@@ -61,32 +58,31 @@ namespace sharpclean
                 {
                     buffer = s.getBuffer();
                     perimeter = s.getPerimeter();
-
                     data[1] = buffer.Count;
                     data[2] = data[1] / s.getEdges();
                     data[0] = getAverageValue(Convert.ToInt32(data[1]));
-
                     conf c = confidence.getconfidence(data);
 
                     if (!c.isObj)
                         colorbuffer(255, Convert.ToInt32(data[1]));
-
+                    
                     if (writeData)
                         printcsv(ref c);
                 }
                 s.clearBuffer();
                 buffer.Clear();
-
                 if (i > per_25 && !b_25)
                 {
                     Console.WriteLine("25%...\n");
                     b_25 = true;
                 }
+
                 if (i > per_50 && !b_50)
                 {
                     Console.WriteLine("50%...\n");
                     b_50 = true;
                 }
+
                 if (i > per_75 && !b_75)
                 {
                     Console.WriteLine("75%...\n");
@@ -98,7 +94,7 @@ namespace sharpclean
         }
 
         //colors a selection of pixels
-        private void colorbuffer(byte color, int sizeofbuffer)
+        private void colorbuffer(int color, int sizeofbuffer)
         {
             for (int i = 0; i < sizeofbuffer; i++)
                 pixels[buffer[i]].value = Convert.ToByte(color);
@@ -115,7 +111,6 @@ namespace sharpclean
         private void printcsv(ref conf c)
         {
             System.IO.File.WriteAllText(ofilename, data[0] + "," + data[1] + "," + data[2] + "," + c.dust + "," + c.obj + ",");
-
             if (c.isObj)
                 System.IO.File.WriteAllText(ofilename, (c.obj - c.dust) + ",obj," + (c.o_val - c.d_val) + "," + (c.o_edge - c.d_edge) + "," + (c.o_size - c.d_size) + "\n");
             else
@@ -127,20 +122,16 @@ namespace sharpclean
         {
             double avg = 0;
             for (int i = 0; i < sizeofbuffer; i++)
-	            avg += pixels[buffer[i]].value;
-
+                avg += pixels[buffer[i]].value;
             return avg / sizeofbuffer;
         }
-        
+
         private pixel[] pixels = null;
         private command cmd = new command();
-
         private int imageWidth, totalPixels;
         private List<int> buffer, perimeter;
         private double[] data = new double[3]; //average value, size, number of edges
-
         private string ofilename;
-
         private readonly string toolbox_err = "::TOOLBOX::error : ";
     }
 }
